requirements.txt
asgiref==3.6.0
Django==3.2.17
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.4.0
